Solidity half of the killed bitcoin lanes, made to compile
The four lanes died mid-edit on a rate limit and left the tree Stack-too-deep
in two places. Both are now fixed and the build is green with no EIP-170
violations. This is the work preserved at tag wip/bitcoin-lanes-2026-09-09
(cf08b451) plus the two stack fixes.

Carries R-vBTC (vBTC is the share token: mint to the LP, transferable, the
fee stream moves with it, redeemVBtc via _resize) and the GATE 3 rulings in
BTCChannels/BitcoinTx (R-P2MR's two enumerated SPK forms, R-LPETH, R-7f).

THE TWO STACK FIXES, both of which deleted something rather than adding:

1. BtcLib.settleBtcLp's second parameter is GONE. It sat unread from
   2026-08-18 behind `address /*lpEth*/` with a note saying the attribution
   it carried is done by the caller. A parameter nothing reads is a
   parameter (rule 23) -- and this one was not free: transferSharesBody
   needs seven arguments on top of its own frame, and with the dead slot it
   did not compile at all. The unread argument WAS the stack budget the vBTC
   transfer path needed. A commented-out parameter is not documentation, it
   is a cost nobody was pricing.

2. transferSharesBody lost its levPooled parameter and the plainNet cap
   moved to Vault.transferShares, which already has the mapping in scope.
   The cap is not relaxed -- it is load-bearing (the levered slice is
   unwind-only and non-transferable, or it walks to a fresh address while
   the venue holds the debt) -- so both sides now say so, and the library
   docblock states that any new caller must repeat it.

3. BitcoinTx's output scanner: sLenBytes is block-scoped because it dies on
   the next line and R-P2MR's packed return (witness version in bits 96+, so
   a caller can tell 0x51 from 0x52 without re-reading the tx) needs the
   slot.

Sizes: SwapLib 24140 (+436), LevManager 24062 (+514), LevMath 24505 (+71),
no violations.

NOT included: evm/src/Aux.sol is dirty with a peer session's in-flight edit
(they added a revert InvalidParam() and the import, while I was building).
Left entirely alone.

No new tests, per owner. forge test NOT run.

+++ evm/src/BTCChannels.sol
@@ -2270,8 +2270,10 @@ contract BTCChannels {
         // §SESS-114 — DUST, not replay. The revert still unwinds the used-mark (state rolls back
         // with it), so the mechanics were always right; only the name was wrong.
         if (sats == 0) revert SwapOutDust();
-        // uint96 packing self-evidently safe (BTC supply ≪ 2^96; 6-dec usd ≪ 2^96/1e6).
-        if (sats > type(uint96).max || usd6 > type(uint96).max) revert InvalidParam();
+        // (R-7f / §BTC-2.5a-quinquies) EACH GUARD IS THE WIDTH THE FIELD IS STORED AT — `sats`
+        // is written `uint64(sats)` below, `usd` is `uint96(usd6)`. Guarding `sats` at uint96 let
+        // (2^64, 2^96] pass and truncate on store; the reject is now at the door and loud.
+        if (sats > type(uint64).max || usd6 > type(uint96).max) revert InvalidParam();
         pendingOnchainSwapOut[swapId] = PendingOnchainSwapOut({
             swapper:           msg.sender,
             sats:              uint64(sats),
